refactor: 초기화 로직 Initializer로 분리
