+++ 3-period/java-subway-path-precourse-main/src/main/java/subway/Application.java
@@ -2,6 +2,7 @@ package subway;
 
 import java.util.Scanner;
 import subway.controller.SubwayController;
+import subway.domain.Initializer;
 import subway.exception.ExceptionHandler;
 import subway.service.SubwayService;
 import subway.view.InputView;
@@ -14,7 +15,9 @@ public class Application {
         OutputView outputView = new OutputView();
         ExceptionHandler exceptionHandler = new ExceptionHandler(outputView);
         SubwayService subwayService = new SubwayService();
-        SubwayController controller = new SubwayController(inputView, outputView, exceptionHandler, subwayService);
+        Initializer initializer = new Initializer();
+        SubwayController controller = new SubwayController(inputView, outputView, exceptionHandler, subwayService,
+                initializer);
         try {
             controller.process();
         } finally {
